feat: Update Api, implement Rayon

+++ src/quantum/computer.rs
@@ -11,7 +11,7 @@ pub struct QuantumComputer {
 }
 
 impl QuantumComputer {
-    pub fn new(q_bits: u32, seed: Option<u64>) -> QuantumComputer {
+    pub fn new(q_bits: usize, seed: Option<u64>) -> QuantumComputer {
         QuantumComputer {
             state: Ket::new(q_bits),
             seed: SmallRng::seed_from_u64(seed.unwrap_or(42)),
@@ -31,11 +31,11 @@ impl QuantumComputer {
         for state_id in 0..self.state.vec.len() {
             for bit_id in 0..self.state.size {
                 if state_id & (0x1 << bit_id) > 0 {
-                    #[cfg(not(test))]
+                    #[cfg(all(not(test), not(feature = "safe")))]
                     {
                         *unsafe { probalillites.get_unchecked_mut(bit_id as usize) } += unsafe { self.state.vec.get_unchecked(state_id) }.amplitude();
                     }
-                    #[cfg(test)]
+                    #[cfg(any(test, feature = "safe"))]
                     {
                         *{ probalillites.get_mut(bit_id as usize).unwrap() } += self.state.vec.get(state_id).unwrap().amplitude();
                     }
@@ -46,17 +46,17 @@ impl QuantumComputer {
         return probalillites
     }
 
-    pub fn probability(&self, bit: u32) -> f64 {
+    pub fn probability(&self, bit: usize) -> f64 {
         assert!(self.state.size > bit);
         let mut proballily = 0.0;
         let bit_m = 0x1 << bit;
         for state_id in 0..self.state.vec.len() {
             if state_id & bit_m > 0 {
-                #[cfg(not(test))]
+                #[cfg(all(not(test), not(feature = "safe")))]
                 {
                     proballily += unsafe { self.state.vec.get_unchecked(state_id) }.amplitude();
                 }
-                #[cfg(test)]
+                #[cfg(any(test, feature = "safe"))]
                 {
                     proballily += self.state.vec.get(state_id).unwrap().amplitude();
                 }
