+++ src/quantum/operator/rotation/rotation_hadamard.rs
@@ -38,12 +38,12 @@ use tsify::Tsify;
 #[cfg_attr(feature = "wasm-pack", serde(rename_all = "camelCase"))]
 #[derive(Copy, Clone, PartialEq)]
 pub struct RotationHadamard {
-    wire: u32,
+    wire: usize,
     theta: f64,
 }
 
 impl RotationHadamard {
-    pub fn new(wire: u32, theta: f64) -> RotationHadamard {
+    pub fn new(wire: usize, theta: f64) -> RotationHadamard {
         RotationHadamard { wire, theta }
     }
 }
@@ -55,7 +55,7 @@ impl ToGate<1> for RotationHadamard {
 }
 
 impl UsedWires<1> for RotationHadamard {
-    fn wires(&self) -> [u32; 1] {
+    fn wires(&self) -> [usize; 1] {
         [self.wire]
     }
 }
