+++ src/quantum/operator/rotation/rotation_z.rs
@@ -28,12 +28,12 @@ use tsify::Tsify;
 #[cfg_attr(feature = "wasm-pack", serde(rename_all = "camelCase"))]
 #[derive(Copy, Clone, PartialEq)]
 pub struct RotationZ {
-    wire: u32,
+    wire: usize,
     theta: f64,
 }
 
 impl RotationZ {
-    pub fn new(wire: u32, theta: f64) -> RotationZ {
+    pub fn new(wire: usize, theta: f64) -> RotationZ {
         RotationZ { wire, theta }
     }
 }
@@ -45,7 +45,7 @@ impl ToGate<1> for RotationZ {
 }
 
 impl UsedWires<1> for RotationZ {
-    fn wires(&self) -> [u32; 1] {
+    fn wires(&self) -> [usize; 1] {
         [self.wire]
     }
 }
