+++ src/quantum/operator/simple/swap.rs
@@ -40,12 +40,12 @@ use tsify::Tsify;
 #[cfg_attr(feature = "wasm-pack", serde(rename_all = "camelCase"))]
 #[derive(Copy, Clone, PartialEq)]
 pub struct Swap {
-    wire_0: u32,
-    wire_1: u32,
+    wire_0: usize,
+    wire_1: usize,
 }
 
 impl Swap {
-    pub fn new(wire_0: u32, wire_1: u32) -> Swap {
+    pub fn new(wire_0: usize, wire_1: usize) -> Swap {
         Swap { wire_0, wire_1 }
     }
 }
@@ -63,7 +63,7 @@ impl ApplyGate<2> for Swap {
 }
 
 impl UsedWires<2> for Swap {
-    fn wires(&self) -> [u32; 2] {
+    fn wires(&self) -> [usize; 2] {
         [self.wire_1, self.wire_0]
     }
 }
