+++ src/quantum/operator/simple/swap_root.rs
@@ -40,12 +40,12 @@ use tsify::Tsify;
 #[cfg_attr(feature = "wasm-pack", serde(rename_all = "camelCase"))]
 #[derive(Copy, Clone, PartialEq)]
 pub struct SwapRoot {
-    wire_0: u32,
-    wire_1: u32,
+    wire_0: usize,
+    wire_1: usize,
 }
 
 impl SwapRoot {
-    pub fn new(wire_0: u32, wire_1: u32) -> SwapRoot {
+    pub fn new(wire_0: usize, wire_1: usize) -> SwapRoot {
         SwapRoot { wire_0, wire_1 }
     }
 }
@@ -63,7 +63,7 @@ impl ApplyGate<2> for SwapRoot {
 }
 
 impl UsedWires<2> for SwapRoot {
-    fn wires(&self) -> [u32; 2] {
+    fn wires(&self) -> [usize; 2] {
         [self.wire_1, self.wire_0]
     }
 }
